Move HTTP client functionality to `BambooHR.Client.Req`
