+++ lib/bamboo_hr/client.ex
@@ -73,7 +73,7 @@ defmodule BambooHR.Client do
   """
   @spec get_company_information(config()) :: response()
   def get_company_information(config) do
-    get("/company_information", config)
+    client_module().get("/company_information", config)
   end
 
   @doc """
@@ -93,7 +93,7 @@ defmodule BambooHR.Client do
   """
   @spec get_company_eins(config()) :: response()
   def get_company_eins(config) do
-    get("/company_eins", config)
+    client_module().get("/company_eins", config)
   end
 
   @doc """
@@ -116,7 +116,7 @@ defmodule BambooHR.Client do
   """
   @spec get_employee(config(), integer(), list(String.t())) :: response()
   def get_employee(config, employee_id, fields) when is_integer(employee_id) do
-    get("/employees/#{employee_id}", config, params: [fields: Enum.join(fields, ",")])
+    client_module().get("/employees/#{employee_id}", config, params: [fields: Enum.join(fields, ",")])
   end
 
   @doc """
@@ -135,7 +135,7 @@ defmodule BambooHR.Client do
   """
   @spec add_employee(config(), map()) :: response()
   def add_employee(config, employee_data) do
-    post("/employees", config, json: employee_data)
+    client_module().post("/employees", config, json: employee_data)
   end
 
   @doc """
@@ -155,7 +155,7 @@ defmodule BambooHR.Client do
   """
   @spec update_employee(config(), integer(), map()) :: response()
   def update_employee(config, employee_id, employee_data) when is_integer(employee_id) do
-    post("/employees/#{employee_id}", config, json: employee_data)
+    client_module().post("/employees/#{employee_id}", config, json: employee_data)
   end
 
   @doc """
@@ -179,7 +179,7 @@ defmodule BambooHR.Client do
   """
   @spec get_employee_directory(config()) :: response()
   def get_employee_directory(config) do
-    get("/employees/directory", config)
+    client_module().get("/employees/directory", config)
   end
 
   @doc """
@@ -211,7 +211,7 @@ defmodule BambooHR.Client do
   """
   @spec get_timesheet_entries(config(), map()) :: response()
   def get_timesheet_entries(config, params) do
-    get("/time_tracking/timesheet_entries", config, params: params)
+    client_module().get("/time_tracking/timesheet_entries", config, params: params)
   end
 
   @doc """
@@ -237,7 +237,7 @@ defmodule BambooHR.Client do
   """
   @spec store_timesheet_clock_entries(config(), list(map())) :: response()
   def store_timesheet_clock_entries(config, entries) do
-    post("/time_tracking/clock_entries/store", config, json: %{items: entries})
+    client_module().post("/time_tracking/clock_entries/store", config, json: %{items: entries})
   end
 
   @doc """
@@ -261,7 +261,7 @@ defmodule BambooHR.Client do
   """
   @spec clock_in_employee(config(), integer(), map()) :: response()
   def clock_in_employee(config, employee_id, clock_data) when is_integer(employee_id) do
-    post("/time_tracking/employees/#{employee_id}/clock_in", config, json: clock_data)
+    client_module().post("/time_tracking/employees/#{employee_id}/clock_in", config, json: clock_data)
   end
 
   @doc """
@@ -285,45 +285,8 @@ defmodule BambooHR.Client do
   """
   @spec clock_out_employee(config(), integer(), map()) :: response()
   def clock_out_employee(config, employee_id, clock_data) when is_integer(employee_id) do
-    post("/time_tracking/employees/#{employee_id}/clock_out", config, json: clock_data)
+    client_module().post("/time_tracking/employees/#{employee_id}/clock_out", config, json: clock_data)
   end
 
-  defp get(path, config, opts \\ []) do
-    request(:get, path, config, opts)
-  end
-
-  defp post(path, config, opts) do
-    request(:post, path, config, opts)
-  end
-
-  defp request(method, path, config, opts) do
-    url = build_url(config, path)
-    headers = build_headers(config.api_key)
-
-    req_opts = Keyword.merge([headers: headers], opts)
-
-    case Req.new(url: url)
-         |> Req.merge(req_opts)
-         |> Req.request(method: method) do
-      {:ok, %{status: status, body: body}} when status in 200..299 ->
-        {:ok, body}
-
-      {:ok, %{status: status, body: body}} ->
-        {:error, %{status: status, body: body}}
-
-      {:error, error} ->
-        {:error, error}
-    end
-  end
-
-  defp build_url(config, path) do
-    "#{config.base_url}/#{config.company_domain}/v1#{path}"
-  end
-
-  defp build_headers(api_key) do
-    [
-      {"Authorization", "Basic " <> Base.encode64("#{api_key}:x")},
-      {"Accept", "application/json"}
-    ]
-  end
+  def client_module, do: Application.get_env(:bamboo_hr, :http_client, BambooHR.Client.Req)
 end
